# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "مستندات API برای ثبت‌نام و احراز هویت کاربران",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/profile": {
            "get": {
                "tags": [
                    "User (User)"
                ],
                "summary": "دریافت پروفایل کاربر (User)",
                "operationId": "27277bb6e0e7dadbce5593fe08a78552",
                "responses": {
                    "200": {
                        "description": "اطلاعات پروفایل کاربر"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User (User)"
                ],
                "summary": "ویرایش پروفایل کاربر (User)",
                "operationId": "12cb2fc71e337236bcd659eedaf17406",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "پروفایل با موفقیت بروزرسانی شد"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/change-password": {
            "put": {
                "tags": [
                    "User (User)"
                ],
                "summary": "تغییر رمز عبور کاربر (User)",
                "operationId": "46c480f1020268e938cc8f5850034760",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "رمز عبور با موفقیت تغییر کرد"
                    },
                    "400": {
                        "description": "رمز عبور فعلی اشتباه است"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/referral-stats": {
            "get": {
                "tags": [
                    "User (User)"
                ],
                "summary": "دریافت آمار ارجاع‌های کاربر (User)",
                "operationId": "6568756e7a243600fbd625acaec2d28b",
                "responses": {
                    "200": {
                        "description": "آمار ارجاع‌ها"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/referral-history": {
            "get": {
                "tags": [
                    "User (User)"
                ],
                "summary": "دریافت تاریخچه تراکنش‌های ارجاع کاربر (User)",
                "operationId": "dcaed18ce0a131b4a974b598adb55977",
                "responses": {
                    "200": {
                        "description": "تاریخچه تراکنش‌های ارجاع"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}